Move AI operations from ToolbarExtension into AIExtension
handleGenerate and handleExtractEntities (plus their helpers
$appendTokenBeforeCaret, $removeAICaret, WHITESPACE_RE) now live
in AIExtension where they belong. The toolbar just calls them
via useAI().

AIExtension now declares dependencies on AICaretNodeExtension and
entity node extensions since it creates those nodes. ToolbarExtension
is reduced to pure toolbar state (block type, formatting, undo/redo).
